Lineup-ClickOnTeam
When clicking on a team (Program.vue) on the program you get to the Lineup page matching the program id.

If it's an announcemnet or a boardingschool, the lineup is set to false, because they don't have a lineup of equipment.

In this version, all lineup data is dummy and hardcoded. So the dynamic data is needed in the next one